requirements.txt
fastapi
uvicorn
asyncpg
starlette
python-dotenv
